Refactor: cleaning code and directory
