+++ src/app/(frontend)/(dashboard)/components/ui/SearchFilter.tsx
@@ -1,6 +1,6 @@
 // src\app\(dashboard)\components\ui\SearchFilter.tsx
 import React from "react";
-import { Search, Filter } from "lucide-react";
+import { Search } from "lucide-react";
 
 interface FilterOption {
   value: string;
