+++ src/app/(frontend)/(dashboard)/items/components/ItemMasterModal.tsx
@@ -125,7 +125,7 @@ export const ItemMasterModal: React.FC<ItemMasterModalProps> = ({
         setErrors({ submit: result.error || "Terjadi kesalahan" });
       }
     } catch (error) {
-      setErrors({ submit: "Terjadi kesalahan tidak terduga" });
+      setErrors({ submit: `Terjadi kesalahan tidak terduga error: ${error}` });
     } finally {
       setIsSubmitting(false);
     }
