+++ src/app/(frontend)/(dashboard)/items/components/ItemModal.tsx
@@ -211,7 +211,7 @@ export const ItemModal: React.FC<ItemModalProps> = ({
         setErrors({ submit: result.error || "Terjadi kesalahan" });
       }
     } catch (error) {
-      setErrors({ submit: "Terjadi kesalahan tidak terduga" });
+      setErrors({ submit: `Terjadi kesalahan tidak terduga error: ${error}` });
     } finally {
       setIsSubmitting(false);
     }
